build: support -Wconversion warning
Some compound assignments called on non-standard trivial types (e.g,
short int) were split into operators and assignments because of strange
gcc behavior (https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56908).

+++ src/Client/UnixPlainStream.hpp
@@ -86,7 +86,7 @@ create_msghdr(struct iovec *iov, size_t iov_count)
 {
 	struct msghdr msg{};
 	msg.msg_iov = iov;
-	msg.msg_iovlen = iov_count;
+	msg.msg_iovlen = static_cast<int>(iov_count);
 	return msg;
 }
 } // namespace internal
